@@ -29,7 +29,7 @@
     <a class="dismiss">🗙</a>
 </div>
 
-<script src="_framework/blazor.server.js"></script>
 <script src="_content/MudBlazor/MudBlazor.min.js"></script>
+<script src="_framework/blazor.server.js"></script>
 </body>
 </html>
