@@ -1,19 +1,30 @@
-using System.Threading;
+using System.Text.Json.Serialization;
 using IOT_WateringSensor;
 using IOT_WateringSensor.Database;
 using IOT_WateringSensor.MQTT_Gøj;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
+using IOT_WateringSensorHub.Areas.Identity;
+using Microsoft.AspNetCore.Components.Authorization;
+using Microsoft.AspNetCore.Identity;
 using MudBlazor.Services;
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
 builder.Services.ConfigureServices();
+builder.Services.AddControllers()
+    .AddJsonOptions(options =>
+    {
+        options.JsonSerializerOptions.Converters.Add(new JsonStringEnumConverter());
+        options.JsonSerializerOptions.PropertyNamingPolicy = null;
+    });
 
+builder.Services.AddDatabaseDeveloperPageExceptionFilter();
+builder.Services.AddDefaultIdentity<IdentityUser>(options => options.SignIn.RequireConfirmedAccount = true)
+    .AddEntityFrameworkStores<WaterSensorDbContext>();
 builder.Services.AddRazorPages();
 builder.Services.AddServerSideBlazor();
 builder.Services.AddMudServices();
+builder.Services.AddScoped<AuthenticationStateProvider, RevalidatingIdentityAuthenticationStateProvider<IdentityUser>>();
+
 new Thread(async _ =>
 {
     var serviceCollection = builder.Services.BuildServiceProvider();
@@ -31,13 +42,31 @@ if (!app.Environment.IsDevelopment())
     //app.UseHsts();
 }
 
-
+app.UseHttpsRedirection();
 
 app.UseStaticFiles();
 
 app.UseRouting();
 
-app.MapBlazorHub();
-app.MapFallbackToPage("/_Host");
+app.UseAuthentication();
+app.UseAuthorization();
+
+app.MapWhen(ctx => !ctx.Request.Path.StartsWithSegments("/api"), blazor =>
+{
+    blazor.UseRouting();
+    blazor.UseEndpoints(endpoints =>
+    {
+        endpoints.MapFallbackToPage("/_Host");
+    });
+});
+
+//explicitly map api endpoints only when path starts with api
+app.MapWhen(ctx => ctx.Request.Path.StartsWithSegments("/api"), api =>
+{
+    api.UseEndpoints(endpoints =>
+    {
+        endpoints.MapControllers();
+    });
+});
 
 app.Run();
